chore: further break down components
